fix tasks interactor
