--- app/interactors/tasks/destroy_organize.rb
@@ -1,8 +1,0 @@
-module Tasks
-  class Destroy
-    include Interactor::Organizer
-
-    organize Tasks::Destroy::Execute,
-             Tasks::Destroy::SendNotification
-  end
-end
