+++ app/interactors/tasks/update/send_notification.rb
@@ -3,9 +3,10 @@ module Tasks
     class SendNotification
       include Interactor
 
-      delegate :project, :task, to: :context
+      delegate :task, to: :context
 
       def call
+        context.project = task.project
         TaskMailer.task_updated(project, task).deliver_later
       end
     end
