--- app/interactors/tasks/update_organize.rb
@@ -1,8 +1,0 @@
-module Tasks
-  class Update
-    include Interactor::Organizer
-
-    organize Tasks::Update::Execute,
-             Tasks::Update::SendNotification
-  end
-end
